backend/src/training/requirements-sagemaker.txt
# Python dependencies for SageMaker YOLOv8 training

# Core dependencies
boto3==1.34.0
sagemaker==2.199.0

# YOLOv8 and dependencies
ultralytics==8.0.196
torch>=2.1.0
torchvision>=0.16.0
opencv-python-headless==4.8.1.78

# Image processing
Pillow>=10.0.0
numpy>=1.24.0

# Data handling
PyYAML>=6.0
pandas>=2.0.0

# Progress tracking
tqdm>=4.65.0

# Visualization (for training plots)
matplotlib>=3.7.0
seaborn>=0.12.0
